optin BL implemented
