+++ ts/features/fastLogin/store/selectors/index.ts
@@ -5,12 +5,15 @@ import { fastLoginEnabled } from "../../../../config";
 import { GlobalState } from "../../../../store/reducers/types";
 import { isPropertyWithMinAppVersionEnabled } from "./../../../../store/reducers/backendStatus";
 
+const fastLoginOptInSelector = (state: GlobalState) =>
+  state.features.loginFeatures.fastLogin.optIn;
+
 /**
  * return the remote config about FastLogin enabled/disabled
  * based on a minumum version of the app.
  * if there is no data, false is the default value -> (FastLogin disabled)
  */
-export const isFastLoginEnabledSelector = createSelector(
+export const isFastLoginFFEnabled = createSelector(
   backendStatusSelector,
   backendStatus =>
     isPropertyWithMinAppVersionEnabled(
@@ -20,18 +23,37 @@ export const isFastLoginEnabledSelector = createSelector(
     )
 );
 
-export const fastLoginSelector = (state: GlobalState) =>
-  state.features.loginFeatures.fastLogin;
+/**
+ * return the remote config about FastLogin enabled/disabled
+ * based on a minumum version of the app, combined with FL opt-in user preference.
+ * If there is no data from backed nor user expressed an optin preference,
+ * false is the default value -> (FastLogin disabled)
+ */
+export const isFastLoginEnabledSelector = createSelector(
+  backendStatusSelector,
+  fastLoginOptInSelector,
+  (backendStatus, optIn) =>
+    isPropertyWithMinAppVersionEnabled(
+      fastLoginEnabled,
+      "fastLogin",
+      backendStatus
+    ) && optIn.fastLoginEnabled
+);
+
+export const fastLoginTokenRefreshHandlerSelector = (state: GlobalState) =>
+  state.features.loginFeatures.fastLogin.tokenRefreshHandler;
 
 export const fastLoginPendingActionsSelector = createSelector(
-  fastLoginSelector,
-  fastLoginState => uniqWith(fastLoginState.pendingActions, isEqual)
+  fastLoginTokenRefreshHandlerSelector,
+  fastLoginTokenRefreshHandlerSelector =>
+    uniqWith(fastLoginTokenRefreshHandlerSelector.pendingActions, isEqual)
 );
 
 export const isFastLoginUserInteractionNeededForSessionExpiredSelector = (
   state: GlobalState
 ) =>
-  state.features.loginFeatures.fastLogin.userInteractionForSessionExpiredNeeded;
+  state.features.loginFeatures.fastLogin.tokenRefreshHandler
+    .userInteractionForSessionExpiredNeeded;
 
 export const tokenRefreshSelector = (state: GlobalState) =>
-  state.features.loginFeatures.fastLogin.tokenRefresh;
+  state.features.loginFeatures.fastLogin.tokenRefreshHandler.tokenRefresh;
