+++ ts/store/actions/types.ts
@@ -25,9 +25,9 @@ import { PnActions } from "../../features/pn/store/actions";
 import { FciActions } from "../../features/fci/store/actions";
 import { IDPayActions } from "../../features/idpay/common/actions";
 import { LollipopActions } from "../../features/lollipop/store/actions/lollipop";
-import { FastLoginActions } from "../../features/fastLogin/store/actions";
 import { WhatsNewActions } from "../../features/whatsnew/store/actions";
 import { WalletV3Actions } from "../../features/walletV3/common/store/actions";
+import { FastLoginActions } from "../../features/fastLogin/store/actions";
 import { AnalyticsActions } from "./analytics";
 import { ApplicationActions } from "./application";
 import { AuthenticationActions } from "./authentication";
